requirements.txt
Arpeggio==1.9.2
attrs==19.3.0
boto3==1.10.18
botocore==1.13.18
dill==0.3.1.1
Django==2.2.13
django-environ==0.4.5
django-storages==1.7.2
django-tinymce==2.8.0
docutils==0.15.2
enum34==1.1.6
futures==3.1.1
jmespath==0.9.4
parver==0.2.1
Pillow==7.1.0
protobuf==3.10.0
psycopg2-binary==2.8.6
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
semver==2.9.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
